requirements.txt
django>=2.1,<2.2
django-cors-headers>=2.4,<2.5
django-debug-toolbar>=1.11,<1.12
djangorestframework>=3.9,<3.10
factory_boy>=2.11,<2.12
